+++ src/utils.js
@@ -306,23 +306,6 @@ module.exports = {
   },
   updateTimestamp: updateTimestamp,
   Promise: _Promise,
-  deepFreeze: function deepFreeze(o) {
-    if (typeof Object.freeze === 'function' && typeof Object.isFrozen === 'function') {
-      var prop, propKey;
-      Object.freeze(o); // First freeze the object.
-      for (propKey in o) {
-        prop = o[propKey];
-        if (!prop || !o.hasOwnProperty(propKey) || typeof prop !== 'object' || Object.isFrozen(prop)) {
-          // If the object is on the prototype, not an object, or is already frozen,
-          // skip it. Note that this might leave an unfrozen reference somewhere in the
-          // object if there is an already frozen object containing an unfrozen object.
-          continue;
-        }
-
-        deepFreeze(prop); // Recursively call deepFreeze.
-      }
-    }
-  },
   compute: function (fn, field, DSUtils) {
     var _this = this;
     var args = [];